Fix shell separation of concerns.
