+++ packages/sdk/shell/src/panels/IdentityPanel/steps/index.ts
@@ -2,7 +2,7 @@
 // Copyright 2023 DXOS.org
 //
 
-export * from './AgentPanel';
+export * from './AgentForm';
 export * from './DeviceManager';
 export * from './IdentityActionChooser';
 export * from './ProfileForm';
